Migrates to new operator-based country identification
MCC are not unique in regards to a given country / territory, they can be assigned to multiple ones.
This commit makes use of the advances in the generate_operatory.py script which now adds further details to each operator.
Thus, CellGuard can identify the correct country name and ISO code for each operator based on its MCC & MNC.